Rename some files, fix hover effect on header, center footer text
