+++ copies/asset-copies/manifest copy.json
