+++ src/assets/css/app.css
@@ -7,10 +7,6 @@
 .in-body{
   color: white;
 }
-.to-head{
-  color: white;
-  align-content: flex-start;
-}
 .header{
 color: white;
 background-color: black;
