+++ src/components/LiveScores.js
@@ -1,6 +1,6 @@
 import React from "react";
 import Header from "./Header";
-import Footer from "./footer";
+import Footer from "./Footer";
 
 
 function LiveScores() {
