+++ src/pages/GBSpredictorinfo.js
@@ -1,6 +1,6 @@
 import React from "react";
 import Header from "./../components/Header";
-import Footer from "./../components/footer";
+import Footer from "../components/Footer";
 import './../assets/css/GBS.css'
