+++ src/pages/golfbettingsites.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './../assets/css/app.css'
-import Footer from '../components/footer';
+import Footer from '../components/Footer';
 import Header from './../components/Header';
 import {useState} from 'react';
 import StarRating from './../components/stars';
