+++ src/pages/livegolfodds.js
@@ -3,7 +3,7 @@ import './../assets/css/app.css';
 import './../assets/css/index.css';
 import Header from './../components/Header';
 import MultiVariableSlider from './../components/sliders';
-import Footer from './../components/footer';
+import Footer from '../components/Footer';
 
 //function NewPage
